+++ brownie-config.yml
@@ -1,7 +1,7 @@
 # use Ganache's forked mainnet mode as the default network (mainnet-fork)
 # NOTE: You don't *have* to do this, but it is often helpful for testing
 networks:
-  default: mainnet-fork
+  default: mainnet-anvil-fork
 
 # automatically fetch contract sources from Etherscan
 autofetch_sources: True
