+++ contracts/interfaces/IVault.sol
@@ -1,0 +1,243 @@
+// SPDX-License-Identifier: GPL-3.0
+pragma solidity >=0.8.18;
+
+import {IERC4626} from "@openzeppelin/contracts/interfaces/IERC4626.sol";
+
+interface IVault is IERC4626 {
+    // STRATEGY EVENTS
+    event StrategyChanged(address indexed strategy, uint256 change_type);
+    event StrategyReported(
+        address indexed strategy,
+        uint256 gain,
+        uint256 loss,
+        uint256 current_debt,
+        uint256 protocol_fees,
+        uint256 total_fees,
+        uint256 total_refunds
+    );
+    // DEBT MANAGEMENT EVENTS
+    event DebtUpdated(
+        address indexed strategy,
+        uint256 current_debt,
+        uint256 new_debt
+    );
+    // ROLE UPDATES
+    event RoleSet(address indexed account, uint256 role);
+    event UpdateRoleManager(address indexed role_manager);
+
+    event UpdateAccountant(address indexed accountant);
+    event UpdateDefaultQueue(address[] new_default_queue);
+    event UpdateUseDefaultQueue(bool use_default_queue);
+    event UpdatedMaxDebtForStrategy(
+        address indexed sender,
+        address indexed strategy,
+        uint256 new_debt
+    );
+    event UpdateDepositLimit(uint256 deposit_limit);
+    event UpdateMinimumTotalIdle(uint256 minimum_total_idle);
+    event UpdateProfitMaxUnlockTime(uint256 profit_max_unlock_time);
+    event DebtPurchased(address indexed strategy, uint256 amount);
+    event Shutdown();
+
+    struct StrategyParams {
+        uint256 activation;
+        uint256 last_report;
+        uint256 current_debt;
+        uint256 max_debt;
+    }
+
+    function FACTORY() external view returns (uint256);
+
+    function strategies(address) external view returns (StrategyParams memory);
+
+    function default_queue(uint256) external view returns (address);
+
+    function use_default_queue() external view returns (bool);
+
+    function minimum_total_idle() external view returns (uint256);
+
+    function deposit_limit() external view returns (uint256);
+
+    function deposit_limit_module() external view returns (address);
+
+    function withdraw_limit_module() external view returns (address);
+
+    function accountant() external view returns (address);
+
+    function roles(address) external view returns (uint256);
+
+    function role_manager() external view returns (address);
+
+    function future_role_manager() external view returns (address);
+
+    function isShutdown() external view returns (bool);
+
+    function nonces(address) external view returns (uint256);
+
+    function initialize(
+        address,
+        string memory,
+        string memory,
+        address,
+        uint256
+    ) external;
+
+    function set_accountant(address new_accountant) external;
+
+    function set_default_queue(address[] memory new_default_queue) external;
+
+    function set_use_default_queue(bool) external;
+
+    function set_deposit_limit(uint256 deposit_limit) external;
+
+    function set_deposit_limit(uint256 deposit_limit, bool should_override)
+        external;
+
+    function set_deposit_limit_module(address new_deposit_limit_module)
+        external;
+
+    function set_deposit_limit_module(
+        address new_deposit_limit_module,
+        bool should_override
+    ) external;
+
+    function set_withdraw_limit_module(address new_withdraw_limit_module)
+        external;
+
+    function set_minimum_total_idle(uint256 minimum_total_idle) external;
+
+    function setProfitMaxUnlockTime(uint256 new_profit_max_unlock_time)
+        external;
+
+    function set_role(address account, uint256 role) external;
+
+    function add_role(address account, uint256 role) external;
+
+    function remove_role(address account, uint256 role) external;
+
+    function transfer_role_manager(address role_manager) external;
+
+    function accept_role_manager() external;
+
+    function unlockedShares() external view returns (uint256);
+
+    function pricePerShare() external view returns (uint256);
+
+    function get_default_queue() external view returns (address[] memory);
+
+    function process_report(address strategy)
+        external
+        returns (uint256, uint256);
+
+    function buy_debt(address strategy, uint256 amount) external;
+
+    function add_strategy(address new_strategy) external;
+
+    function revoke_strategy(address strategy) external;
+
+    function force_revoke_strategy(address strategy) external;
+
+    function update_max_debt_for_strategy(
+        address strategy,
+        uint256 new_max_debt
+    ) external;
+
+    function update_debt(address strategy, uint256 target_debt)
+        external
+        returns (uint256);
+
+    function update_debt(
+        address strategy,
+        uint256 target_debt,
+        uint256 max_loss
+    ) external returns (uint256);
+
+    function shutdown_vault() external;
+
+    function totalIdle() external view returns (uint256);
+
+    function totalDebt() external view returns (uint256);
+
+    function apiVersion() external view returns (string memory);
+
+    function assess_share_of_unrealised_losses(
+        address strategy,
+        uint256 assets_needed
+    ) external view returns (uint256);
+
+    function profitMaxUnlockTime() external view returns (uint256);
+
+    function fullProfitUnlockDate() external view returns (uint256);
+
+    function profitUnlockingRate() external view returns (uint256);
+
+    function lastProfitUpdate() external view returns (uint256);
+
+    //// NON-STANDARD ERC-4626 FUNCTIONS \\\\
+
+    function withdraw(
+        uint256 assets,
+        address receiver,
+        address owner,
+        uint256 max_loss
+    ) external returns (uint256);
+
+    function withdraw(
+        uint256 assets,
+        address receiver,
+        address owner,
+        uint256 max_loss,
+        address[] memory strategies
+    ) external returns (uint256);
+
+    function redeem(
+        uint256 shares,
+        address receiver,
+        address owner,
+        uint256 max_loss
+    ) external returns (uint256);
+
+    function redeem(
+        uint256 shares,
+        address receiver,
+        address owner,
+        uint256 max_loss,
+        address[] memory strategies
+    ) external returns (uint256);
+
+    function maxWithdraw(address owner, uint256 max_loss)
+        external
+        view
+        returns (uint256);
+
+    function maxWithdraw(
+        address owner,
+        uint256 max_loss,
+        address[] memory strategies
+    ) external view returns (uint256);
+
+    function maxRedeem(address owner, uint256 max_loss)
+        external
+        view
+        returns (uint256);
+
+    function maxRedeem(
+        address owner,
+        uint256 max_loss,
+        address[] memory strategies
+    ) external view returns (uint256);
+
+    //// NON-STANDARD ERC-20 FUNCTIONS \\\\
+
+    function DOMAIN_SEPARATOR() external view returns (bytes32);
+
+    function permit(
+        address owner,
+        address spender,
+        uint256 amount,
+        uint256 deadline,
+        uint8 v,
+        bytes32 r,
+        bytes32 s
+    ) external returns (bool);
+}
