+++ tests/conftest.py
@@ -17,6 +17,7 @@ chain_used = 1
 
 ################################################## TENDERLY DEBUGGING ##################################################
 
+
 # change autouse to True if we want to use this fork to help debug tests
 @pytest.fixture(scope="session", autouse=use_tenderly)
 def tenderly_fork(web3, chain):
@@ -36,19 +37,28 @@ def tenderly_fork(web3, chain):
 #################### FIXTURES BELOW NEED TO BE ADJUSTED FOR THIS REPO ####################
 
 
+# for this, important to use a vault that is currently failing with the old strategy version (ie, 1 wei that stucks around)
 @pytest.fixture(scope="session")
-def token():
-    token_address = "0xc4AD29ba4B3c580e6D59105FFf484999997675Ff"  # this should be the address of the ERC-20 used by the strategy/vault (curve 3Crypto)
+def token(use_v3):
+    if use_v3:
+        token_address = "0x6B175474E89094C44Da98b954EedeAC495271d0F"  # DAI
+    else:
+        token_address = "0xD2967f45c4f384DEEa880F807Be904762a3DeA07"  # this should be the address of the ERC-20 used by the strategy/vault (curve GUSD-3CRV)
     yield interface.IERC20(token_address)
 
 
 @pytest.fixture(scope="session")
-def whale(amount, token):
+def whale(amount, token, use_v3):
     # Totally in it for the tech
     # Update this with a large holder of your want token (the largest EOA holder of LP)
-    whale = accounts.at(
-        "0x347140c7F001452e6A60131D24b37103D0e34231", force=True
-    )  # 0x347140c7F001452e6A60131D24b37103D0e34231, 3crypto pool, 127 tokens
+    if use_v3:
+        whale = accounts.at(
+            "0xD1668fB5F690C59Ab4B0CAbAd0f8C1617895052B", force=True
+        )  # 0xD1668fB5F690C59Ab4B0CAbAd0f8C1617895052B, EOA, 47M DAI
+    else:
+        whale = accounts.at(
+            "0xB41742195962ca2D9886690AC2854aBf7B826090", force=True
+        )  # 0xB41742195962ca2D9886690AC2854aBf7B826090, EOA, 737k tokens
     if token.balanceOf(whale) < 2 * amount:
         raise ValueError(
             "Our whale needs more funds. Find another whale or reduce your amount variable."
@@ -58,16 +68,21 @@ def whale(amount, token):
 
 @pytest.fixture(scope="session")
 def amount(token):
-    amount = 50 * 10 ** token.decimals()
+    amount = 50_000 * 10 ** token.decimals()
     yield amount
 
 
 @pytest.fixture(scope="session")
-def profit_whale(profit_amount, token):
+def profit_whale(profit_amount, token, use_v3):
     # ideally not the same whale as the main whale, or else they will lose money
-    profit_whale = accounts.at(
-        "0xc7599b60f05639f93D26e58d56D90C526A6e7575", force=True
-    )  # 0xc7599b60f05639f93D26e58d56D90C526A6e7575, 3crypto pool, 105 tokens
+    if use_v3:
+        profit_whale = accounts.at(
+            "0x837c20D568Dfcd35E74E5CC0B8030f9Cebe10A28", force=True
+        )  # 0x837c20D568Dfcd35E74E5CC0B8030f9Cebe10A28, eoa, 34M DAI
+    else:
+        profit_whale = accounts.at(
+            "0xdE86EA8867E63a4133E255a3E2FF912dfeA6D658", force=True
+        )  # 0xdE86EA8867E63a4133E255a3E2FF912dfeA6D658, eoa, 230k tokens
     if token.balanceOf(profit_whale) < 5 * profit_amount:
         raise ValueError(
             "Our profit whale needs more funds. Find another whale or reduce your profit_amount variable."
@@ -77,35 +92,60 @@ def profit_whale(profit_amount, token):
 
 @pytest.fixture(scope="session")
 def profit_amount(token):
-    profit_amount = 1 * 10 ** token.decimals()
+    profit_amount = 100 * 10 ** token.decimals()
     yield profit_amount
 
 
 # set address if already deployed, use ZERO_ADDRESS if not
 @pytest.fixture(scope="session")
-def vault_address():
-    vault_address = "0xE537B5cc158EB71037D4125BDD7538421981E6AA"
+def vault_address(use_v3):
+    if use_v3:
+        vault_address = "0xdA816459F1AB5631232FE5e97a05BBBb94970c95"
+    else:
+        vault_address = "0x2a38B9B0201Ca39B17B460eD2f11e4929559071E"
     yield vault_address
 
 
 # if our vault is pre-0.4.3, this will affect a few things
 @pytest.fixture(scope="session")
-def old_vault():
-    old_vault = True
+def old_vault(use_v3):
+    if use_v3:
+        old_vault = False
+    else:
+        old_vault = True
     yield old_vault
 
 
 # this is the name we want to give our strategy
 @pytest.fixture(scope="session")
-def strategy_name():
-    strategy_name = "StrategyRouterV2"
+def strategy_name(use_v3):
+    if use_v3:
+        strategy_name = "StrategyRouterV3"
+    else:
+        strategy_name = "StrategyRouterV2"
     yield strategy_name
 
 
 # this is the name of our strategy in the .sol file
 @pytest.fixture(scope="session")
-def contract_name(StrategyRouterV2):
-    contract_name = StrategyRouterV2
+def contract_name(
+    StrategyRouterV2,
+    StrategyRouterV3,
+    V3Router,
+    StrategyRouterV2Old,
+    use_old,
+    use_v3,
+):
+    if use_v3:
+        if use_old:
+            contract_name = V3Router
+        else:
+            contract_name = StrategyRouterV3
+    else:
+        if use_old:
+            contract_name = StrategyRouterV2Old
+        else:
+            contract_name = StrategyRouterV2
     yield contract_name
 
 
@@ -273,14 +313,23 @@ def strategy(
     strategy = gov.deploy(contract_name, vault, destination_vault, strategy_name)
 
     strategy.setKeeper(keeper, {"from": gov})
-    strategy.setHealthCheck(health_check, {"from": gov})
-    strategy.setDoHealthCheck(True, {"from": gov})
-    vault.setPerformanceFee(0, {"from": gov})
-    vault.setManagementFee(0, {"from": gov})
+    # strategy.setHealthCheck(health_check, {"from": gov})
+    # strategy.setDoHealthCheck(True, {"from": gov})
+    if vault.performanceFee() != 0:
+        vault.setPerformanceFee(0, {"from": gov})
+    if vault.managementFee() != 0:
+        vault.setManagementFee(0, {"from": gov})
+
+    # migrate to our new strategy to inherit the broken state
+    old_router = vault.withdrawalQueue(0)
+    vault.migrateStrategy(old_router, strategy, {"from": gov})
+
+    # turn on health check for first harvest since we're inheriting profit
+    # strategy.setDoHealthCheck(False, {"from": gov})
 
     # if we have other strategies, set them to zero DR and remove them from the queue
     if vault_address != ZERO_ADDRESS:
-        for i in range(0, 20):
+        for i in range(1, 20):
             strat_address = vault.withdrawalQueue(i)
             if ZERO_ADDRESS == strat_address:
                 break
@@ -289,8 +338,6 @@ def strategy(
                 vault.updateStrategyDebtRatio(strat_address, 0, {"from": gov})
                 interface.ICurveStrategy045(strat_address).harvest({"from": gov})
                 vault.removeStrategyFromQueue(strat_address, {"from": gov})
-
-    vault.addStrategy(strategy, 10_000, 0, 2**256 - 1, 0, {"from": gov})
 
     # turn our oracle into testing mode by setting the provider to 0x00, then forcing true
     strategy.setBaseFeeOracle(base_fee_oracle, {"from": management})
@@ -306,6 +353,26 @@ def strategy(
 ####################         PUT UNIQUE FIXTURES FOR THIS REPO BELOW         ####################
 
 
+# tried parameterizing these two but brownie did not seem to like it and started locking up
+# use this for whether we want to test the old version of the strategy
+@pytest.fixture(scope="session")
+def use_old():
+    yield False
+
+
+# use this if we're doing a V2 or V3 router
+# NOTE: we've only updated testing for simple_harvest to work with the V3 strategies
+@pytest.fixture(scope="session")
+def use_v3():
+    yield True
+
+
+# flag to denote if we're migrating from existing strategies and thus will likely have profit on our first harvest
+@pytest.fixture(scope="session")
+def is_migration():
+    yield True
+
+
 # use this similarly to how we use use_yswaps
 @pytest.fixture(scope="session")
 def is_gmx():
@@ -319,12 +386,18 @@ def is_router():
 
 
 @pytest.fixture(scope="session")
-def destination_vault(interface):
-    # destination vault of the route
-    yield interface.IVaultFactory045("0x8078198Fc424986ae89Ce4a910Fc109587b6aBF3")
+def destination_vault(use_v3):
+    # destination vault of the route.
+    if use_v3:
+        yield Contract("0x028eC7330ff87667b6dfb0D94b954c820195336c")
+    else:
+        yield interface.IVaultFactory045("0x63bD3Bbb6c5cb6E457C3f3cbb2D8aa2536E319F1")
 
 
 @pytest.fixture(scope="session")
-def destination_strategy():
-    # destination strategy of the route
-    yield interface.ICurveStrategy045("0x9D7CD0041ABd91f281E282Db3fba7A9Db9E4cC8b")
+def destination_strategy(destination_vault, use_v3):
+    # destination curve strategy of the route
+    if use_v3:
+        yield Contract("0xAeDF7d5F3112552E110e5f9D08c9997Adce0b78d")
+    else:
+        yield interface.ICurveStrategy045(destination_vault.withdrawalQueue(1))
