+++ tests/test_odds_and_ends_part_1.py
@@ -3,6 +3,7 @@ from brownie import Contract, chain, ZERO_ADDRESS, interface
 import pytest
 from utils import harvest_strategy, check_status
 
+
 # this module includes other tests we may need to generate, for instance to get best possible coverage on prepareReturn or liquidatePosition
 # do any extra testing here to hit all parts of liquidatePosition
 # generally this involves sending away all assets and then withdrawing before another harvest
@@ -21,32 +22,40 @@ def test_liquidatePosition(
     RELATIVE_APPROX,
     vault_address,
     is_gmx,
+    use_v3,
     destination_vault,
     is_router,
+    is_migration,
 ):
     ## deposit to the vault after approving
     starting_whale = token.balanceOf(whale)
     token.approve(vault, 2**256 - 1, {"from": whale})
     vault.deposit(amount, {"from": whale})
     (profit, loss, extra) = harvest_strategy(
-        use_yswaps,
+        use_v3,
         strategy,
         token,
         gov,
         profit_whale,
         profit_amount,
         target,
+        destination_vault,
     )
 
     # check our current status
     print("\nAfter first harvest")
     strategy_params = check_status(strategy, vault)
 
+    # simulate 5 days of waiting for share price to bump back up
+    chain.sleep(86400 * 5)
+    chain.mine(1)
+
     # evaluate our current total assets
     old_assets = vault.totalAssets()
     initial_debt = strategy_params["totalDebt"]
     starting_share_price = vault.pricePerShare()
     initial_strategy_assets = strategy.estimatedTotalAssets()
+    loose_want = token.balanceOf(vault)
 
     ################# SEND ALL FUNDS AWAY. ADJUST AS NEEDED PER STRATEGY. #################
     to_send = destination_vault.balanceOf(strategy)
@@ -60,8 +69,12 @@ def test_liquidatePosition(
     # we shouldn't have taken any actual losses yet, and debt/DR/share price should all still be the same
     assert strategy_params["debtRatio"] == 10_000
     assert strategy_params["totalLoss"] == 0
-    assert strategy_params["totalDebt"] == initial_debt == old_assets
-    assert vault.pricePerShare() == starting_share_price
+    if is_migration:
+        assert strategy_params["totalDebt"] == initial_debt == old_assets - loose_want
+        assert vault.pricePerShare() >= starting_share_price
+    else:
+        assert strategy_params["totalDebt"] == initial_debt == old_assets
+        assert vault.pricePerShare() == starting_share_price
 
     # if slippery, then assets may differ slightly from debt
     if is_slippery:
@@ -96,13 +109,16 @@ def test_liquidatePosition(
     if vault_address == ZERO_ADDRESS:
         # because we only withdrew, and didn't harvest, only the withdrawn portion of assets will be known as a loss
         # however, if we own all of the vault, this is a moot point
-        assert strategy_params["totalGain"] == 0
+        if is_migration:
+            assert strategy_params["totalGain"] == profit
+        else:
+            assert strategy_params["totalGain"] == 0
+            assert vault.pricePerShare() == 10 ** token.decimals()
 
         # if our asset conversion has some slippage, then we may still have a few wei of assets left in the vault.
         # not accounting for that in this testing but may be worth adding in the future, especially if some of the asserts below fail.
 
         # if we burn all vault shares, then price per share is 1.0
-        assert vault.pricePerShare() == 10 ** token.decimals()
 
         # everything is gone 😅
         if is_gmx:
@@ -123,7 +139,13 @@ def test_liquidatePosition(
         if not is_router:
             # router won't have taken any loss at all since this is a larger-than-dust sized loss
             assert strategy_params["totalLoss"] > 0
-        assert strategy_params["totalGain"] == 0
+
+        # worth noting that while we still had locked profit during the withdrawal above, got some weird results
+        # with price per share where share price actually did go down—worth looking into someday maybe?
+        if is_migration:
+            assert strategy_params["totalGain"] == profit
+        else:
+            assert strategy_params["totalGain"] == 0
 
         # share price isn't affected since shares are burned proportionally with losses
         assert vault.pricePerShare() == starting_share_price
@@ -180,21 +202,26 @@ def test_locked_funds(
     target,
     use_yswaps,
     is_gmx,
+    use_v3,
     destination_vault,
 ):
+    if use_v3:
+        return
+
     # should update this one for Router
     ## deposit to the vault after approving
     starting_whale = token.balanceOf(whale)
     token.approve(vault, 2**256 - 1, {"from": whale})
     vault.deposit(amount, {"from": whale})
     (profit, loss, extra) = harvest_strategy(
-        use_yswaps,
+        use_v3,
         strategy,
         token,
         gov,
         profit_whale,
         profit_amount,
         target,
+        destination_vault,
     )
 
     # check our current status
@@ -209,22 +236,24 @@ def test_locked_funds(
     chain.sleep(sleep_time)
 
     # send away our funds from our destination strategy
-    voter = target.curveVoter()
+    voter = Contract(target.proxy()).proxy()
     gauge = Contract(target.gauge())
     to_send = gauge.balanceOf(voter)
     print("Gauge Balance of Vault", to_send)
-    gauge.transfer(gov, to_send, {"from": voter})
+    gauge.withdraw(to_send, {"from": voter})
+    token.transfer(gov, to_send, {"from": voter})
     assert target.estimatedTotalAssets() == 0
 
     strategy.setDoHealthCheck(False, {"from": gov})
     (profit, loss, extra) = harvest_strategy(
-        use_yswaps,
+        use_v3,
         strategy,
         token,
         gov,
         profit_whale,
         0,
         target,
+        destination_vault,
     )
 
     print("Loss:", loss, "\nInitial Deposit:", amount)
@@ -247,21 +276,26 @@ def test_locked_funds_zero_dr(
     target,
     use_yswaps,
     is_gmx,
+    use_v3,
     destination_vault,
 ):
+    if use_v3:
+        return
+
     # should update this one for Router
     ## deposit to the vault after approving
     starting_whale = token.balanceOf(whale)
     token.approve(vault, 2**256 - 1, {"from": whale})
     vault.deposit(amount, {"from": whale})
     (profit, loss, extra) = harvest_strategy(
-        use_yswaps,
+        use_v3,
         strategy,
         token,
         gov,
         profit_whale,
         profit_amount,
         target,
+        destination_vault,
     )
 
     # check our current status
@@ -276,11 +310,12 @@ def test_locked_funds_zero_dr(
     chain.sleep(sleep_time)
 
     # send away our funds from our destination strategy
-    voter = target.curveVoter()
+    voter = Contract(target.proxy()).proxy()
     gauge = Contract(target.gauge())
     to_send = gauge.balanceOf(voter)
     print("Gauge Balance of Vault", to_send)
-    gauge.transfer(gov, to_send, {"from": voter})
+    gauge.withdraw(to_send, {"from": voter})
+    token.transfer(gov, to_send, {"from": voter})
     assert target.estimatedTotalAssets() == 0
 
     # set our debtRatio to zero to return all debt
@@ -289,13 +324,14 @@ def test_locked_funds_zero_dr(
 
     strategy.setDoHealthCheck(False, {"from": gov})
     (profit, loss, extra) = harvest_strategy(
-        use_yswaps,
+        use_v3,
         strategy,
         token,
         gov,
         profit_whale,
         0,
         target,
+        destination_vault,
     )
 
     print("Loss:", loss, "\nInitial Deposit:", amount)
@@ -315,22 +351,25 @@ def test_rekt(
     use_yswaps,
     old_vault,
     is_gmx,
+    use_v3,
     destination_vault,
     is_slippery,
     RELATIVE_APPROX,
+    is_migration,
 ):
     ## deposit to the vault after approving
     starting_whale = token.balanceOf(whale)
     token.approve(vault, 2**256 - 1, {"from": whale})
     vault.deposit(amount, {"from": whale})
-    (profit, loss, extra) = harvest_strategy(
-        use_yswaps,
+    (first_profit, loss, extra) = harvest_strategy(
+        use_v3,
         strategy,
         token,
         gov,
         profit_whale,
         profit_amount,
         target,
+        destination_vault,
     )
 
     # check our current status
@@ -351,20 +390,22 @@ def test_rekt(
     assert strategy.estimatedTotalAssets() == 0
 
     # our whale donates 5 wei to the vault so we don't divide by zero (needed for older vaults)
-    if old_vault:
+    # if we migrate we'll have some profit sitting in our vault so not a big deal
+    if old_vault and not is_migration:
         token.transfer(strategy, 5, {"from": whale})
 
     # set debtRatio to zero so we try and pull everything that we can out. turn off health check because of massive losses
     vault.updateStrategyDebtRatio(strategy, 0, {"from": gov})
     strategy.setDoHealthCheck(False, {"from": gov})
     (profit, loss, extra) = harvest_strategy(
-        use_yswaps,
+        use_v3,
         strategy,
         token,
         gov,
         profit_whale,
         profit_amount,
         target,
+        destination_vault,
     )
     if is_gmx:
         assert strategy.estimatedTotalAssets() == extra
@@ -374,7 +415,10 @@ def test_rekt(
         assert strategy.estimatedTotalAssets() == 0
 
     if old_vault:
-        assert vault.totalAssets() == 5
+        if is_migration:
+            assert vault.totalAssets() == first_profit
+        else:
+            assert vault.totalAssets() == 5
     else:
         assert vault.totalAssets() == 0
 
@@ -428,342 +472,3 @@ def test_weird_reverts(
     # can't withdraw from a non-vault address
     with brownie.reverts():
         strategy.withdraw(1e18, {"from": gov})
-
-
-# this test makes sure we can still harvest without any assets but still get our profits
-# can also test here whether we claim rewards from an empty strategy, some protocols will revert
-def test_empty_strat(
-    gov,
-    token,
-    vault,
-    whale,
-    strategy,
-    amount,
-    profit_whale,
-    profit_amount,
-    target,
-    use_yswaps,
-    old_vault,
-    is_slippery,
-    RELATIVE_APPROX,
-    vault_address,
-    sleep_time,
-    is_gmx,
-    destination_vault,
-):
-    ## deposit to the vault after approving
-    starting_whale = token.balanceOf(whale)
-    token.approve(vault, 2**256 - 1, {"from": whale})
-    vault.deposit(amount, {"from": whale})
-    (profit, loss, extra) = harvest_strategy(
-        use_yswaps,
-        strategy,
-        token,
-        gov,
-        profit_whale,
-        profit_amount,
-        target,
-    )
-
-    # check our current status
-    print("\nAfter first harvest")
-    strategy_params = check_status(strategy, vault)
-
-    # evaluate our current total assets
-    old_assets = vault.totalAssets()
-    initial_strategy_assets = strategy.estimatedTotalAssets()
-    initial_debt = strategy_params["totalDebt"]
-    starting_share_price = vault.pricePerShare()
-
-    # sleep to get some yield
-    chain.sleep(sleep_time)
-
-    ################# SEND ALL FUNDS AWAY. ADJUST AS NEEDED PER STRATEGY. #################
-    to_send = destination_vault.balanceOf(strategy)
-    destination_vault.transfer(gov, to_send, {"from": strategy})
-
-    # confirm we emptied the strategy
-    assert strategy.estimatedTotalAssets() == 0
-
-    # check that our losses are approximately the whole strategy
-    print("\nBefore dust transfer, after main fund transfer")
-    strategy_params = check_status(strategy, vault)
-
-    # we shouldn't have taken any actual losses yet, and debt/DR/share price should all still be the same
-    assert strategy_params["debtRatio"] == 10_000
-    assert strategy_params["totalLoss"] == 0
-    assert strategy_params["totalDebt"] == initial_debt == old_assets
-    assert vault.pricePerShare() == starting_share_price
-
-    # if slippery, then assets may differ slightly from debt
-    if is_slippery:
-        assert (
-            pytest.approx(initial_debt, rel=RELATIVE_APPROX) == initial_strategy_assets
-        )
-    else:
-        assert initial_debt == initial_strategy_assets
-
-    # confirm we emptied the strategy
-    assert strategy.estimatedTotalAssets() == 0
-
-    # our whale donates 5 wei to the vault so we don't divide by zero (needed for older vaults)
-    # old vaults also don't have the totalIdle var
-    if old_vault:
-        dust_donation = 5
-        token.transfer(strategy, dust_donation, {"from": whale})
-        assert strategy.estimatedTotalAssets() == dust_donation
-    else:
-        total_idle = vault.totalIdle()
-        assert total_idle == 0
-
-    # check our current status
-    print("\nBefore harvest, after funds transfer out + dust transfer in")
-    strategy_params = check_status(strategy, vault)
-
-    # we shouldn't have taken any actual losses yet, and debt/DR/share price should all still be the same
-    assert strategy_params["debtRatio"] == 10_000
-    assert strategy_params["totalLoss"] == 0
-    assert strategy_params["totalDebt"] == initial_debt == old_assets
-    assert vault.pricePerShare() == starting_share_price
-    assert vault.debtOutstanding(strategy) == 0
-
-    # accept our losses, sad day 🥲
-    strategy.setDoHealthCheck(False, {"from": gov})
-    (profit, loss, extra) = harvest_strategy(
-        use_yswaps,
-        strategy,
-        token,
-        gov,
-        profit_whale,
-        profit_amount,
-        target,
-    )
-    assert loss > 0
-
-    # check our status
-    print("\nAfter our big time loss")
-    strategy_params = check_status(strategy, vault)
-
-    # DR goes to zero, loss is > 0, gain and debt should be near zero (zero for new vaults), share price also nearr zero (bye-bye assets 💀)
-    if old_vault:
-        assert strategy_params["debtRatio"] == 1
-    else:
-        assert strategy_params["debtRatio"] == 0
-    assert strategy_params["totalLoss"] > 0
-    if not is_gmx:
-        assert strategy_params["totalGain"] == 0
-        assert vault.pricePerShare() == 0
-
-    # vault should also have no assets, except old ones will have 5 wei
-    if old_vault:
-        assert strategy_params["totalDebt"] == dust_donation == vault.totalAssets()
-        assert strategy.estimatedTotalAssets() <= dust_donation
-    else:
-        assert strategy_params["totalDebt"] == 0 == vault.totalAssets()
-        total_idle = vault.totalIdle()
-        assert total_idle == 0
-        if use_yswaps:
-            assert strategy.estimatedTotalAssets() == profit_amount
-        elif is_gmx:
-            assert strategy.estimatedTotalAssets() == extra
-        else:
-            assert strategy.estimatedTotalAssets() == 0
-
-    print("Total supply:", vault.totalSupply())
-
-    # some profits fall from the heavens
-    # this should be used to pay down debt vs taking profits
-    token.transfer(strategy, profit_amount, {"from": profit_whale})
-    print("\nAfter our donation")
-    strategy_params = check_status(strategy, vault)
-    strategy.setDoHealthCheck(False, {"from": gov})
-    (profit, loss, extra) = harvest_strategy(
-        use_yswaps,
-        strategy,
-        token,
-        gov,
-        profit_whale,
-        profit_amount,
-        target,
-    )
-    assert profit > 0
-    share_price = vault.pricePerShare()
-    assert share_price > 0
-    print("Share price:", share_price)
-
-
-# this test makes sure we can still harvest without any profit and not revert
-# for some strategies it may be impossible to harvest without generating profit, especially if not using yswaps
-def test_no_profit(
-    gov,
-    token,
-    vault,
-    whale,
-    strategy,
-    amount,
-    profit_whale,
-    profit_amount,
-    target,
-    use_yswaps,
-    sleep_time,
-    is_gmx,
-):
-    ## deposit to the vault after approving
-    token.approve(vault, 2**256 - 1, {"from": whale})
-    vault.deposit(amount, {"from": whale})
-    (profit, loss, extra) = harvest_strategy(
-        use_yswaps,
-        strategy,
-        token,
-        gov,
-        profit_whale,
-        profit_amount,
-        target,
-    )
-
-    # check our current status
-    print("\nAfter first harvest")
-    strategy_params = check_status(strategy, vault)
-
-    # store our starting share price
-    starting_share_price = vault.pricePerShare()
-
-    # normally we would sleep here, but we are intentionally trying to avoid profit, so we don't
-
-    # if are using yswaps and we don't want profit, don't use yswaps (False for first argument).
-    # Or just don't harvest our destination strategy, can pass 0 for profit_amount and use if statement in utils
-    (profit, loss, extra) = harvest_strategy(
-        False,
-        strategy,
-        token,
-        gov,
-        profit_whale,
-        0,
-        target,
-    )
-
-    # check our current status
-    print("\nAfter harvest")
-    strategy_params = check_status(strategy, vault)
-
-    assert profit == 0
-    assert vault.pricePerShare() == starting_share_price
-
-
-# test some gmx-specific functions
-def test_gmx_vesting(
-    gov,
-    token,
-    vault,
-    whale,
-    strategy,
-    amount,
-    sleep_time,
-    is_slippery,
-    no_profit,
-    profit_whale,
-    profit_amount,
-    target,
-    use_yswaps,
-    is_gmx,
-):
-    if not is_gmx:
-        return
-    elif to_vest == 0:
-        strategy.rebalanceVesting({"from": gov})
-        return
-
-    # rebalance when we have nothing
-    strategy.rebalanceVesting({"from": gov})
-
-    ## deposit to the vault after approving
-    starting_whale = token.balanceOf(whale)
-    token.approve(vault, 2**256 - 1, {"from": whale})
-    vault.deposit(amount, {"from": whale})
-
-    # harvest, store asset amount
-    (profit, loss, extra) = harvest_strategy(
-        use_yswaps,
-        strategy,
-        token,
-        gov,
-        profit_whale,
-        profit_amount,
-        target,
-    )
-    old_assets = vault.totalAssets()
-    assert old_assets > 0
-    assert strategy.estimatedTotalAssets() > 0
-
-    # simulate profits
-    chain.sleep(sleep_time)
-
-    # harvest, store new asset amount
-    (profit, loss, extra) = harvest_strategy(
-        use_yswaps,
-        strategy,
-        token,
-        gov,
-        profit_whale,
-        profit_amount,
-        target,
-    )
-
-    # make sure we are vesting
-    assert strategy.vestingEsMpx() > 0
-
-    # rebalance vesting
-    strategy.rebalanceVesting({"from": gov})
-    assert strategy.vestingEsMpx() > 0
-
-    # simulate profits
-    chain.sleep(sleep_time)
-
-    # harvest, store new asset amount
-    (profit, loss, extra) = harvest_strategy(
-        use_yswaps,
-        strategy,
-        token,
-        gov,
-        profit_whale,
-        profit_amount,
-        target,
-    )
-
-    # simulate profits
-    chain.sleep(sleep_time)
-
-    # harvest, store new asset amount
-    (profit, loss, extra) = harvest_strategy(
-        use_yswaps,
-        strategy,
-        token,
-        gov,
-        profit_whale,
-        profit_amount,
-        target,
-    )
-
-    # simulate profits
-    chain.sleep(sleep_time)
-
-    # harvest, store new asset amount
-    (profit, loss, extra) = harvest_strategy(
-        use_yswaps,
-        strategy,
-        token,
-        gov,
-        profit_whale,
-        profit_amount,
-        target,
-    )
-
-    # we shouldn't be able to sweep out more than we have, but it also won't revert
-    mpx = interface.IERC20(strategy.mpx())
-    before = mpx.balanceOf(gov)
-    strategy.unstakeAndSweepVestedMpx(strategy.stakedMpx() * 2, {"from": gov})
-    assert before == mpx.balanceOf(gov)
-
-    # sweep out our MPX
-    strategy.unstakeAndSweepVestedMpx(strategy.stakedMpx(), {"from": gov})
-    assert mpx.balanceOf(gov) > 0
