+++ tests/test_simple_harvest.py
@@ -2,6 +2,7 @@ from brownie import chain, Contract
 from utils import harvest_strategy
 import pytest
 
+
 # test the our strategy's ability to deposit, harvest, and withdraw, with different optimal deposit tokens if we have them
 def test_simple_harvest(
     gov,
@@ -18,6 +19,8 @@ def test_simple_harvest(
     target,
     use_yswaps,
     is_gmx,
+    use_v3,
+    destination_vault,
 ):
     ## deposit to the vault after approving
     starting_whale = token.balanceOf(whale)
@@ -25,15 +28,18 @@ def test_simple_harvest(
     vault.deposit(amount, {"from": whale})
     newWhale = token.balanceOf(whale)
 
+    print("Deposited to vault from whale")
+
     # harvest, store asset amount
     (profit, loss, extra) = harvest_strategy(
-        use_yswaps,
+        use_v3,
         strategy,
         token,
         gov,
         profit_whale,
         profit_amount,
         target,
+        destination_vault,
     )
     old_assets = vault.totalAssets()
     assert old_assets > 0
@@ -44,29 +50,62 @@ def test_simple_harvest(
 
     # harvest, store new asset amount
     (profit, loss, extra) = harvest_strategy(
-        use_yswaps,
+        use_v3,
         strategy,
         token,
         gov,
         profit_whale,
         profit_amount,
         target,
+        destination_vault,
     )
     # record this here so it isn't affected if we donate via ySwaps
     strategy_assets = strategy.estimatedTotalAssets()
+    print("Profit from normal harvest:", profit)
+    print("Loss from normal harvest:", loss)
+    print("Profit from normal harvest with decimals:", profit / 1e18)
 
     # harvest again so the strategy reports the profit
     if use_yswaps or is_gmx:
         print("Using ySwaps for harvests")
         (profit, loss, extra) = harvest_strategy(
-            use_yswaps,
+            use_v3,
             strategy,
             token,
             gov,
             profit_whale,
             profit_amount,
             target,
+            destination_vault,
         )
+
+    # do some checks for bugs from our old router version
+    # harvesting with no profit would fail in old version eventually (takes 1 wei profit)
+    (profit, loss, extra) = harvest_strategy(
+        use_v3,
+        strategy,
+        token,
+        gov,
+        profit_whale,
+        0,
+        target,
+        destination_vault,
+    )
+    print("Profit from no profit harvest #1:", profit)
+    print("Loss from harvest:", loss)
+
+    (profit, loss, extra) = harvest_strategy(
+        use_v3,
+        strategy,
+        token,
+        gov,
+        profit_whale,
+        0,
+        target,
+        destination_vault,
+    )
+    print("Profit from no profit harvest #2:", profit)
+    print("Loss from harvest:", loss)
 
     # evaluate our current total assets
     new_assets = vault.totalAssets()
