+++ tests/test_yswaps.py
@@ -2,6 +2,7 @@ import brownie
 from brownie import ZERO_ADDRESS, interface, chain
 from utils import harvest_strategy
 
+
 # test our permissionless swaps and our trade handler functions as intended
 def test_keepers_and_trade_handler(
     gov,
